Also use node provider in session manager
