+++ etc/omero.properties
@@ -671,7 +671,7 @@ omero.throttling.method_time.error.indexer=86400000
 #############################################
 omero.cluster.redirector=nullRedirector
 omero.cluster.read_only=false
-omero.cluster.node_provider=ome.services.blitz.fire.NodeProvider
+omero.cluster.node_provider=ome.security.basic.BasicNodeProvider
 
 #############################################
 ## grid configuration
